Revert to having stack navigator as a root;
